report-generator/backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
Pillow==10.2.0

# RAG and embeddings
chromadb==0.4.22
sentence-transformers==2.3.1
langchain==0.1.4
langchain-anthropic==0.1.1

# LLM client
anthropic==0.18.1
python-dotenv==1.0.0

# Calculations
numpy==1.26.3
pandas==2.2.0
scipy==1.12.0

# Utilities
tqdm==4.66.1
pyyaml==6.0.1